[FIX] Set expiry for site transient
Change Details
------------
The site transient had no expiry date, which meant that if no file changes were detecetd by `scandir($mudir)` then the old key would continue to be used.

I have set this to have an expiry of 1 day so that if this does happen eventually new code will be loaded, but with a 1 day expiry the performance impact is minimal.

An example of this happening is that if a file already exists in `mu-plugins` but doesn't have the appropriate **plugin name** comment then the file is included in the scandir results, but not loaded by WP because of the missing comment. A later change to add the comment to the file does not prompt the loader to load the file.

Backout Plan
-----------
Revert of this commit will restore previous behaviour without impact.